+++ reptor/plugins/projects/Project/tests/test_integration_project.py
@@ -48,7 +48,7 @@ class TestIntegrationProject(object):
         )
         p.communicate()
         assert p.returncode == 0
-        assert projects_api.fetch_project().readonly is True
+        assert projects_api.get_project().readonly is True
 
         # Reactivate project
         p = subprocess.Popen(
@@ -56,7 +56,7 @@ class TestIntegrationProject(object):
         )
         p.communicate()
         assert p.returncode == 0
-        assert projects_api.fetch_project().readonly is False
+        assert projects_api.get_project().readonly is False
 
     def test_export_tar_gz(self):
         fname = "myproject.tar.gz"